+++ testutil/context.go
@@ -3,15 +3,15 @@ package testutil
 import (
 	"github.com/line/ostracon/libs/log"
 	ocproto "github.com/line/ostracon/proto/ostracon/types"
-	"github.com/line/tm-db/v2/memdb"
+	dbm "github.com/tendermint/tm-db"
 
 	"github.com/line/lbm-sdk/store"
 	sdk "github.com/line/lbm-sdk/types"
 )
 
-// DefaultContext creates a sdk.Context with a fresh MemDB that can be used in tests.
+// DefaultContext creates a sdk.Context with a fresh dbm that can be used in tests.
 func DefaultContext(key sdk.StoreKey, tkey sdk.StoreKey) sdk.Context {
-	db := memdb.NewDB()
+	db := dbm.NewMemDB()
 	cms := store.NewCommitMultiStore(db)
 	cms.MountStoreWithDB(key, sdk.StoreTypeIAVL, db)
 	cms.MountStoreWithDB(tkey, sdk.StoreTypeTransient, db)