+++ x/token/keeper/msg_server.go
@@ -196,6 +196,14 @@ func (s msgServer) GrantPermission(c context.Context, req *token.MsgGrantPermiss
 
 	s.keeper.Grant(ctx, req.ContractId, granter, grantee, permission)
 
+	event := token.EventGrant{
+		ContractId: req.ContractId,
+		Granter:    req.From,
+		Grantee:    req.To,
+		Permission: permission,
+	}
+	ctx.EventManager().EmitEvent(token.NewEventGrantPermToken(event))
+
 	return &token.MsgGrantPermissionResponse{}, nil
 }
 
@@ -213,6 +221,13 @@ func (s msgServer) RevokePermission(c context.Context, req *token.MsgRevokePermi
 	}
 
 	s.keeper.Abandon(ctx, req.ContractId, grantee, permission)
+
+	event := token.EventAbandon{
+		ContractId: req.ContractId,
+		Grantee:    req.From,
+		Permission: permission,
+	}
+	ctx.EventManager().EmitEvent(token.NewEventRevokePermToken(event))
 
 	return &token.MsgRevokePermissionResponse{}, nil
 }